norminetted

+++ ft_strlen.c
@@ -6,7 +6,7 @@
 /*   By: lgrigore <lgrigore@student.42madrid.com    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/01/31 13:20:56 by lgrigore          #+#    #+#             */
-/*   Updated: 2025/01/31 13:21:03 by lgrigore         ###   ########.fr       */
+/*   Updated: 2025/01/31 17:03:49 by lgrigore         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 size_t	ft_strlen(const char *s)
 {
-	int	num;
+	size_t	num;
 
 	num = 0;
 	if (s == NULL)
